+++ mars/tensor/datasource/tests/test_datasource.py
@@ -25,6 +25,7 @@ except (ImportError, OSError):  # pragma: no cover
     tiledb = None
 
 from mars import dataframe as md
+from mars.core import enter_mode
 from mars.tensor import ones, zeros, tensor, full, arange, diag, linspace, triu, tril, ones_like
 from mars.tensor.datasource import array, fromtiledb, TensorTileDBDataSource, fromdense
 from mars.tensor.datasource.tri import TensorTriu, TensorTril
@@ -35,7 +36,6 @@ from mars.tensor.datasource.ones import TensorOnes, TensorOnesLike
 from mars.tensor.core import Tensor, SparseTensor
 from mars.tensor.datasource.from_dataframe import from_dataframe
 from mars.tests.core import TestBase
-from mars.utils import enter_mode
 
 
 class Test(TestBase):
